# openapi.yaml
openapi: 3.0.3
info:
  title: Club Operations POS API
  version: 1.0.0
  description: |
    Server-authoritative REST API for managing club check-ins, room/locker assignments,
    cleaning workflows, checkout verification, and operational reporting.
    
    All endpoints are versioned under `/v1`.
    Request bodies are validated with Zod schemas.
    
    Authentication:
    - Staff endpoints require Bearer token (session token from PIN or WebAuthn login)
    - Admin endpoints require ADMIN role in addition to authentication
    - Customer-facing endpoints (agreement signing, checkout kiosk) do not require auth
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /v1/auth/login-pin:
    post:
      tags: [Authentication]
      summary: Staff login with PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [staffLookup, deviceId, pin]
              properties:
                staffLookup:
                  type: string
                  description: Staff ID or name
                deviceId:
                  type: string
                pin:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  staff:
                    $ref: '#/components/schemas/Staff'
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token from PIN or WebAuthn login
  schemas:
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [STAFF, ADMIN]
        active:
          type: boolean
...